security/manager/ssl/tests/mochitest/bugs/test_bug644006.html

<!DOCTYPE HTML>
<html><head>
  <title>Test bug 644006</title>
  <script type="text/javascript"
    src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css"
    href="chrome://mochikit/content/tests/SimpleTest/test.css" />
</head><body>
<script type="application/javascript;version=1.7">
const FAILURE = Components.results.NS_ERROR_FAILURE;
const ILLEGAL_VALUE = Components.results.NS_ERROR_ILLEGAL_VALUE;

var certDB = Components.classes["@mozilla.org/security/x509certdb;1"]
             .getService(Components.interfaces.nsIX509CertDB);

function excMessage(e)
{
  if ("message" in e && e.message !== null) {
    let msg = e.message;
    if ("data" in e && e.data !== null)
      msg = msg + ": " + e.data;
    return msg;
  } else {
    return e.toString();
  }
}

function testGood(data)
{
  let label = "CN=" + data.cn;
  try {
    let cert = certDB.constructX509FromBase64(data.cert);
    is(cert.commonName, data.cn, label + ": constructX509 succeeded");
  } catch (e) {
    ok(false, label + ": exception: " + excMessage(e));
  }
}

function testBad(data)
{
  let label = uneval(data.t)
  try {
    let cert = certDB.constructX509FromBase64(data.t);
    ok(false, label + ": constructX509 succeeded");
  } catch (e) {
    is(e.result, data.e, label + ": exception: " + excMessage(e));
  }
}

const badCases = [
  // wrong type or too short
  { t: null, e: ILLEGAL_VALUE },
  { t: "",   e: ILLEGAL_VALUE },
  { t: "=",  e: ILLEGAL_VALUE },
  { t: "==", e: ILLEGAL_VALUE },
  // not base64
  { t: "forty-four dead stone lions", e: ILLEGAL_VALUE },
  // not a cert
  { t: "Zm9ydHktZm91ciBkZWFkIHN0b25lIGxpb25z", e: FAILURE }
];

// real certs with all three padding levels
const goodCases = [
  { cn: "A",   cert: "MIHhMIGcAgEAMA0GCSqGSIb3DQEBBQUAMAwxCjAIBgNVBAMTAUEwHhcNMTEwMzIzMjMyNTE3WhcNMTEwNDIyMjMyNTE3WjAMMQowCAYDVQQDEwFBMEwwDQYJKoZIhvcNAQEBBQADOwAwOAIxANFm7ZCfYNJViaDWTFuMClX3+9u18VFGiyLfM6xJrxir4QVtQC7VUC/WUGoBUs9COQIDAQABMA0GCSqGSIb3DQEBBQUAAzEAx2+gIwmuYjJO5SyabqIm4lB1MandHH1HQc0y0tUFshBOMESTzQRPSVwPn77a6R9t" },
  { cn: "Bo",  cert: "MIHjMIGeAgEAMA0GCSqGSIb3DQEBBQUAMA0xCzAJBgNVBAMTAkJvMB4XDTExMDMyMzIzMjYwMloXDTExMDQyMjIzMjYwMlowDTELMAkGA1UEAxMCQm8wTDANBgkqhkiG9w0BAQEFAAM7ADA4AjEA1FoSl9w9HqMqVgk2K0J3OTiRsgHeNsQdPUl6S82ME33gH+E56PcWZA3nse+fpS3NAgMBAAEwDQYJKoZIhvcNAQEFBQADMQAo/e3BvQAmygiATljQ68tWPoWcbMwa1xxAvpWTEc1LOvMqeDBinBUqbAbSmPhGWb4=" },
  { cn: "Cid", cert: "MIHlMIGgAgEAMA0GCSqGSIb3DQEBBQUAMA4xDDAKBgNVBAMTA0NpZDAeFw0xMTAzMjMyMzI2MzJaFw0xMTA0MjIyMzI2MzJaMA4xDDAKBgNVBAMTA0NpZDBMMA0GCSqGSIb3DQEBAQUAAzsAMDgCMQDUUxlF5xKN+8KCSsR83sN+SRwJmZdliXsnBB7PU0OgbmOWN0u8yehRkmu39kN9tzcCAwEAATANBgkqhkiG9w0BAQUFAAMxAJ3UScNqRcjHFrNu4nuwRldZLJlVJvRYXp982V4/kYodQEGN4gJ+Qyj+HTsaXy5x/w==" }
];

var i;
for (i = 0; i < badCases.length; i++)
  testBad(badCases[i]);
for (i = 0; i < goodCases.length; i++)
  testGood(goodCases[i]);
</script>
</body></html>
